GPflow kernel objects

Methods to construct, combine and evaluate kernels (covariance functions)

Usage

 # static kernel objects
 k <- kernels$White(input_dim, variance = 1, active_dims = NULL)
 k <- kernels$Constant(input_dim, variance = 1, active_dims = NULL)
 k <- kernels$Bias(input_dim, variance = 1, active_dims = NULL)
    # stationary kernel objects
 k <- kernels$RBF(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
 k <- kernels$Exponential(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
 k <- kernels$Matern12(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
 k <- kernels$Matern32(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
 k <- kernels$Matern52(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
    # non-stationary kernel objects
 k <- kernels$Linear(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
    # periodic kernel objects
 k <- kernels$Cosine(input_dim, variance = 1, lengthscales = NULL, active_dims = NULL, ARD = FALSE)
 k <- kernels$PeriodicKernel(input_dim, period = 1, variance = 1, lengthscales = NULL, active_dims = NULL)
    # kernel operations
 k + k
 k * k
    # kernel object member functions
 k$K(X, X2 = NULL)
 k$Kdiag(X)
 k$compute_K(X, Z)
 k$compute_K_symm(X)
 #k$compute_Kdiag(X)

Arguments

input_dim

An integer vector giving the dimensions of the matrix X on which member functions will operate.

active_dims

An integer vector identifying the columns of X on which this kernel operates. If NULL (default), all columns are active

variance

A positive numeric scalar giving the initial value of the marginal variance of the kernel.

lengthscales

A positive numeric vector giving the initial value of kernel lengthscales for the columns indexed by active_dims.

period

A positive numeric vector giving the initial value of the periodicity for the columns indexed by active_dims.

ARD

Whether allow lengthscales and periodicities to vary between active dimensions. If FALSE, lengthscales and period should be scalar.