parameter transformation objects

transformation objects used internally in GPflowR

Details

This module will likely be unexported and not formally documented in the near future. This is essentially a test run of the module documentation set up

Usage

 # transform objects
 t <- transforms$Identity()
 t <- transforms$Exp(lower = 1e-6)
 t <- transforms$Log1pe(lower = 1e-6)
 t <- transforms$Logistic(a = 0, b = 1)
 t <- transforms$positive(lower = 1e-6)
    # transform object member functions
 t$forward(x)
 t$backward(y)
 t$tf_forward(x)
 t$tf_log_jacobian(x)
 

Arguments

lower

The minimum value that positive transforms can take. This helps stability during optimization, because aggressive optimizers can take overly-long steps which lead to zero in the transformed variable, causing an error.

a, b

The location (a) and scale (b) of the logistic distribution used in the logistic transform.